dataipa: Fixed RNDIS teardown warning
During RNDIS teardown, there are many warning stack prints taking
place. As a result, the composition switch timing is experiencing
a delay longer than the 3 seconds allocated by the userspace timer
when the composition request begins, causing a timeout with the
userspace.With this change we are now using kernel APIs.

Change-Id: Ibbf840fa1764cf3fa2497bd5a642b1aa6e7e1b44